# backend/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/instagram/webhooks/subscriptions": {
      "post": {
        "operationId": "InstagramWebhooksController_createSubscription",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWebhookSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Subscription created successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create webhook subscription for Instagram account",
        "tags": [
          "Instagram Webhooks"
        ]
      }
    },
    "/api/instagram/webhooks/events/{accountId}": {
      "get": {
        "operationId": "InstagramWebhooksController_getEvents",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eventType",
            "required": false,
            "in": "query",
            "description": "Filter by event type (comment, mention, message, story_mention, live_comment)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "processed",
            "required": false,
            "in": "query",
            "description": "Filter by processed status",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Results per page (default: 50)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Events retrieved successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get webhook events for an Instagram account",
        "tags": [
          "Instagram Webhooks"
        ]
      }
    },
    "/api/instagram/webhooks/stats/{accountId}": {
      "get": {
        "operationId": "InstagramWebhooksController_getStats",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookStatsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get webhook statistics for an Instagram account",
        "tags": [
          "Instagram Webhooks"
        ]
      }
    },
    "/api/instagram/webhooks/retry/{accountId}": {
      "post": {
        "operationId": "InstagramWebhooksController_retryFailedEvents",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Events queued for retry"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Retry failed webhook events",
        "tags": [
          "Instagram Webhooks"
        ]
      }
    },
...
  },
  "info": {
    "title": "Social Selling API",
    "description": "\n# Social Selling Platform API\n\nComplete REST API for Instagram-based social selling platform.\n\n## Features\n- Instagram OAuth integration\n- Product catalog management\n- Direct message automation\n- Post scheduling and analytics\n- Multi-account support\n\n## Authentication\nAll authenticated endpoints require a Bearer token in the Authorization header.\nObtain tokens via POST /api/auth/login endpoint.\n\n## Rate Limiting\n- Authenticated: 1000 requests/hour\n- Unauthenticated: 100 requests/hour\n\n## Session Management\nSessions are managed via HTTP-only cookies and can be tracked per device.\n  ",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local Development"
    },
    {
      "url": "https://staging-api.socialselling.willianbvsanches.com",
      "description": "Staging Environment"
    },
    {
      "url": "https://api.socialselling.willianbvsanches.com",
      "description": "Production Environment"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT access token",
        "in": "header"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ssell_session",
        "description": "Session cookie for authenticated requests"
      }
    },
    "schemas": {
...
      "CreateWebhookSubscriptionDto": {
        "type": "object",
        "properties": {
          "instagramAccountId": {
            "type": "string",
            "description": "Instagram account ID"
          },
          "subscriptionFields": {
            "type": "array",
            "description": "Subscription fields",
            "items": {
              "type": "string",
              "enum": [
                "messages",
                "comments",
                "mentions",
                "story_insights",
                "live_comments"
              ]
            }
          },
          "verifyToken": {
            "type": "string",
            "description": "Custom verify token (auto-generated if not provided)"
          }
        },
        "required": [
          "instagramAccountId",
          "subscriptionFields"
        ]
      },
      "WebhookStatsDto": {
        "type": "object",
        "properties": {
          "totalEvents": {
            "type": "number"
          },
          "processedEvents": {
            "type": "number"
          },
          "pendingEvents": {
            "type": "number"
          },
          "failedEvents": {
            "type": "number"
          },
          "duplicateEvents": {
            "type": "number"
          },
          "eventsByType": {
            "type": "object"
          }
        },
        "required": [
          "totalEvents",
          "processedEvents",
          "pendingEvents",
          "failedEvents",
          "duplicateEvents",
          "eventsByType"
        ]
      },
...
    }
  }
}